pyproject.toml
[project]
name = "pyblish-debug"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

dependencies = [
    "debugpy>=1.8.13",
    "pyblish-base>=1.8.12",
]
